requirements.txt
Django==1.5.4
South==0.8.2
dj-database-url==0.2.2
dj-static==0.0.5
django-dajaxice==0.5.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
